Remove flask from HWI JSONRPC and separate it to another file
